### 2025-07-15 d399704
Fix compilation issues of JOSS paper
