### 2025-07-15 408235e
Add JOSS badge
