### 2025-07-15 #8590 @alastair-marshall
Hi @goerz, I think your recollection is mostly right. So I was interested in helping, had a version in the old QuOptimalControl.jl from my summer holidays working away. I think that eventually I was struggling with implementing things a lot and you had finished Krotov and took over.

I'm not sure whether my contribution warrants inclusion, I did try for a long time, but none of it remains. I would of course love to be included, the package is great!
