### 2025-07-15 #8590 @goerz
I'd err on the side of including all substantial contributions, and I think "early prototyping" absolutely warrants inclusion as an author. I'm sorry I overlooked those early commits when preparing the manuscript!

Can you add your info before [line 15 of the `paper.md`](https://github.com/JuliaQuantumControl/GRAPE.jl/blob/408235ec34bd57870caba2b4afa2614aa9fba85e/paper/paper.md?plain=1#L15) (as well as adding an entry to "affiliations").

See the notes on https://github.com/JuliaQuantumControl/GRAPE.jl/pull/90#issue-3150873070; briefly: you should be able to edit and commit the the file using https://github.dev/JuliaQuantumControl/GRAPE.jl/blob/master/paper/paper.md, if you don't want to do it locally.

@alastair-marshall If possible, can you do it as soon as possible: We have to resubmit some internal paperwork to make this change.
