+++ paper/paper.md
@@ -21,7 +21,7 @@ authors:
 affiliations:
  - name: DEVCOM Army Research Laboratory, 2800 Powder Mill Road, Adelphi, MD 20783, United States
    index: 1
- - name: Universitat Ulm,  Institute for Quantum Optics, Albert-Einstein-Allee 11, 89081 Ulm, Germany
+ - name: Universität Ulm,  Institute for Quantum Optics, Albert-Einstein-Allee 11, 89081 Ulm, Germany
    index: 2
 date: 15 July 2025
 bibliography: paper.bib
@@ -51,7 +51,7 @@ As a compiled language geared towards scientific computing, Julia delivers numer
 
 Julia's core feature of [multiple dispatch](https://www.youtube.com/watch?v=kc9HwsxE1OY) allows the user to define custom, problem-specific data structures with performance-optimized linear algebra operations. This gives `GRAPE.jl` great flexibility to work with any custom data structures for quantum states $\vert \Psi_k(t) \rangle$ or the matrices $\hat{H}_k(\{\epsilon_l(t)\})$, and enables a wide range of applications, from NMR spin systems to superconducting circuits or trapped atoms in quantum computing, to systems with spatial degrees of freedom [e.g., @DashAVSQS2024]. This also includes open quantum systems, as the structure of \autoref{eq:tdse} holds not just for the standard Schrödinger equation, but also for the Liouville equation, where $\vert\Psi_k\rangle$ is replaced by a (vectorized) density matrix and $\hat{H}$ becomes a Liouvillian super-operator [@GoerzNJP2014].
 
-The rise of machine learning generated considerable interest in using the capabilities of frameworks like Tensorflow [@Tensorflow], PyTorch [@PaszkeNIPS2019], or JAX [@JAX] for automatic differentiation (AD) [@Griewank2008] to evaluate the gradient of the optimization functional. This has the benefit that it allows for arbitrary functionals [@LeungPRA2017; @quantum-optimal-control; @AbdelhafezPRA2019; @AbdelhafezPRA2020]. In contrast, the GRAPE method and all of its existing implementations are formulated only for a "standard" set of functionals that essentially measure the overlap of a propagated state with a target state. Unfortunately, AD comes with a large numerical overhead that makes the method impractical. @GoerzQ2022 introduced the use of "semi-automatic differentiation" that limits the numerical cost to exactly that of the traditional GRAPE scheme. It does this by limiting the use of AD to evaluate only the derivative $\partial J/\partial \langle \Psi_k(T) \vert$, and only if that derivative cannot be evaluated analytically. `GRAPE.jl` is built on the resulting generalized GRAPE scheme. As necessary, it can use any available AD framework in the Julia ecosystem to enable the use of non-analytical functionals, such as entanglement measures [@GoerzPRA2015; @WattsPRA2015].
[…]
