### 2025-07-16 5a52727
removing addresses from affils in paper.md
