### 2025-07-16 #8590 @danielskatz
👋 @goerz - I've made a PR (https://github.com/JuliaQuantumControl/GRAPE.jl/pull/91) to remove the addresses from the affiliations.
