### 2025-07-16 e5491d0
Separate bib for paper and documentation
JOSS complained that having the full `.bib` file for the paper was
slowing down the DOI checker too much.
