### 2025-07-16 #8590 @goerz
> Your bib file is very long, with many entries that you aren't using.

Yeah, that's because I was using the same `.bib` file both for the paper and the full documentation.

> Our checker checks all the entries, which takes a long time.

Sorry about that! I've now split the original `.bib` file into two separate `.bib` files, with `paper/paper.bib` only containing those references used `paper/paper.md` and `docs/src/refs.bib` containing everything for the entire documentation.

> So the missing DOI we've possibly found above might or might not be in an entry that you are using in the paper

That one actually was used in the paper

> If it is, please add the DOI, assuming it's correct

Done
