### 2025-08-04 #8590 @goerz
Just bumping this: it would be really wonderful if this could move into the review phase, if there’s any editor available
