+++ .github/workflows/JOSS.yml
@@ -10,7 +10,7 @@ jobs:
     name: Compile Manuscript
     steps:
       - name: Checkout
-        uses: actions/checkout@v4
+        uses: actions/checkout@v5
       - name: Build PDF with Inara (Docker)
         run: |
           docker run --rm \
