+++ .github/workflows/REUSE.yml
@@ -17,7 +17,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
-        uses: actions/checkout@v4
+        uses: actions/checkout@v5
 
       - name: REUSE Compliance Check
         uses: fsfe/reuse-action@v5
