### 2025-08-13 #8590 @Nikoleta-v3
Thank you for your submission, @goerz! I see that you’ve already suggested some reviewers, so I’ll go ahead and invite a few of them. I’m not an expert in quantum at all, but that’s what the reviewers are for.

From an initial look, the package appears to be well documented, and the repository is very clean 👏🏻

In the paper, there’s one typo on line 21: `with → will` or `with be → is`?
