### 2025-08-13 77d3016
Fix typo in JOSS manuscript
As reported in https://github.com/openjournals/joss-reviews/issues/8590#issuecomment-3182097672
