+++ paper/paper.md
@@ -37,7 +37,7 @@ The quantum state of a system can be described numerically by a complex vector $
 i \hbar \frac{\partial \vert\Psi(t)\rangle}{\partial t} = \hat{H}(\epsilon(t)) \vert\Psi(t)\rangle\,,
 \end{equation}
 
-where $\hbar$ is the [reduced Planck constant](https://en.wikipedia.org/wiki/Planck_constant) and $\hat{H}$ is a matrix whose elements depend in some way on the control function $\epsilon(t)$. We generally know the initial state of the system $\vert\Psi(t=0)\rangle$ and want to find an $\epsilon(t)$ that minimizes some real-valued functional $J$ that depends on the states at some final time $T$, as well as running costs on $\vert\Psi(t)\rangle$ and values of $\epsilon(t)$ at intermediate times. A common example with be the square-modulus of the overlap with a target state.
+where $\hbar$ is the [reduced Planck constant](https://en.wikipedia.org/wiki/Planck_constant) and $\hat{H}$ is a matrix whose elements depend in some way on the control function $\epsilon(t)$. We generally know the initial state of the system $\vert\Psi(t=0)\rangle$ and want to find an $\epsilon(t)$ that minimizes some real-valued functional $J$ that depends on the state at some final time $T$, as well as running costs on $\vert\Psi(t)\rangle$ and values of $\epsilon(t)$ at intermediate times. A common example is the square-modulus of the overlap with a target state.
 
 The defining feature of the GRAPE method is that it considers $\epsilon(t)$ as piecewise constant, i.e., as a vector of values $\epsilon_n$, for the $n$'th interval of the time grid. This allows solving \autoref{eq:tdse} for each time interval, and deriving an expression for the gradient $\partial J / \partial \epsilon_n$ of the optimization functional with respect to the values of the control field. It results in an efficient numerical scheme for evaluating the full gradient [@GoerzQ2022, {Figure 1(a)}]. The scheme extends to situations where the functional is evaluated on top of *multiple* propagated states $\{\vert \Psi_k(t) \rangle\}$ with an index $k$, and multiple controls $\epsilon_l(t)$, resulting in a vector of values $\epsilon_{nl}$ with a double-index $nl$. Once the gradient has been evaluated, in the original formulation of GRAPE [@KhanejaJMR2005], the values $\epsilon_{nl}$ would then be updated by taking a step with a fixed step width $\alpha$ in the direction of the negative gradient, to iteratively minimize the value of the optimization functional $J$. In practice, the gradient can also be fed into an arbitrary gradient-based optimizer, and in particular a quasi-Newton method like L-BFGS-B [@ZhuATMS1997; @LBFGSB.jl]. This results in a dramatic improvement in stability and convergence [@FouquieresJMR2011], and is assumed as the default in `GRAPE.jl`. Gradients of the time evolution operator can be evaluated to machine precision following @GoodwinJCP2015. The GRAPE method could also be extended to a true Hessian of the optimization functional [@GoodwinJCP2016], which would be in scope for future versions of `GRAPE.jl`.
 
