### 2025-08-14 #8813 @editorialbot
[REVIEW]: GRAPE.jl: Gradient Ascent Pulse Engineering in Julia
**Submitting author:** @goerz (<a href="http://orcid.org/0000-0003-2839-9976">Michael Goerz</a>)
**Repository:** https://github.com/JuliaQuantumControl/GRAPE.jl
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
