### 2025-09-03 #8813 @Nikoleta-v3
Hey, thanks for the ping @goerz. I understand it’s frustrating waiting on the reviewers. In Europe, “summer holiday” usually refers to August, so I had assumed that by the end of the month most people would be back at work and that we’d have the reviews by mid-September.

It’s of course fine to ask for updates, and to be fair, I probably should have pinged the reviewers a week ago. That being said, let’s keep the tone respectful so we can stay constructive going forward 😄
