### 2025-09-28 #8813 @AlCap23
Sorry for the delayed review @goerz and @Nikoleta-v3 .

Overall the contribution is well written and serves a clear purpose. I agree with @szabo137 review in most aspects. However, I think the limited usage within the QuantumControl Framework is acceptable. A more general framework would be nice, but given that the problem this Algorithm is trying to solve is very specific to Quantum Optimal Control I think it's acceptable.

A minor suggestion would be to make the examples copy-paste able without relying on Pluto. Literate is able to do this. And this would make the examples even more accessible.

On this note:
The tests are very well engineered and use a good set of packages ( SafeTestsSets, StableRNGs ). Very nice 👍.

I installed and tested everything using Julia 1.10.10 and all tests are green.
