### 2025-09-28 #8813 @goerz
@AlCap23 @szabo137 Thank you both for your thoughtful reviews!

I'm in the process of some refactoring of the package API and adding a self-contained, standalone tutorial to the documentation.

We had originally decided to deliberately hide the API of the `GRAPE` package behind the interface of the `QuantumControl` package, even though `GRAPE` could in fact have an independent API, and the packages are more distinct than might immediately be obvious. In hindsight, that seems to have been not the best of decisions to make the package as approachable was we would like it to be. Likewise, the lack of a fully self-contained example in the GRAPE documentation itself is absolutely something that needs to be remedied.

I should have an update and a more detailed response within the next few days.
