### 2025-10-01 ff4085e
Refactor API, add tutorial
* Rename `GRAPE.optimize_grape` to `GRAPE.optimize`
* Clarify relationship to QuantumControl.jl
* Republish `Trajectory` and `set_default_ad_framework`, to ensure
  being self-contained
