+++ src/GRAPE.jl
@@ -7,6 +7,13 @@ module GRAPE
 include("workspace.jl")
 include("result.jl")
 include("optimize.jl")
+include("docstring.jl")
 include(joinpath("..", "ext", "GRAPELBFGSBExt.jl"))
+
+using SciMLPublic: @public
+@public optimize, GrapeResult
+
+using QuantumControl: Trajectory, set_default_ad_framework
+@public Trajectory, set_default_ad_framework
 
 end
