+++ src/docstring.jl
@@ -1,0 +1,192 @@
+# vim: set filetype=markdown :
+#
+# SPDX-FileCopyrightText: © 2025 Michael Goerz <mail@michaelgoerz.net>
+#
+# SPDX-License-Identifier: MIT
+
+@doc raw"""
+Solve a quantum control problem using the GRAPE method.
+
+```julia
+using GRAPE
+result = GRAPE.optimize(trajectories, tlist; J_T, kwargs...)
+```
+
+minimizes a functional
+
+```math
+J(\{ϵ_{nl}\}) = J_T(\{|Ψ_k(T)⟩\}) + λ_a J_a(\{ϵ_{nl}\})\,,
+```
+
+via the GRAPE method, where the final time functional ``J_T`` depends
+explicitly on the forward-propagated states ``|Ψ_k(T)⟩``, where ``|Ψ_k(t)⟩`` is
+the time evolution of the `initial_state` in the ``k``th' element of the
+`trajectories`, and the running cost ``J_a`` depends explicitly on pulse values
+``ϵ_{nl}`` of the l'th control discretized on the n'th interval of the time
+grid `tlist`.
+
+It does this by calculating the gradient of the final-time functional
+
+```math
+\nabla J_T \equiv \frac{\partial J_T}{\partial ϵ_{nl}}
+= -2 \Re
+\underbrace{%
+\underbrace{\bigg\langle χ(T) \bigg\vert \hat{U}^{(k)}_{N_T} \dots \hat{U}^{(k)}_{n+1} \bigg \vert}_{\equiv \bra{\chi(t_n)}\;\text{(bw. prop.)}}
+\frac{\partial \hat{U}^{(k)}_n}{\partial ϵ_{nl}}
+}_{\equiv \bra{χ_k^\prime(t_{n-1})}}
+\underbrace{\bigg \vert \hat{U}^{(k)}_{n-1} \dots \hat{U}^{(k)}_1 \bigg\vert Ψ_k(t=0) \bigg\rangle}_{\equiv |\Psi(t_{n-1})⟩\;\text{(fw. prop.)}}\,,
+```
+
+where ``\hat{U}^{(k)}_n`` is the time evolution operator for the ``n`` the
+interval, generally assumed to be ``\hat{U}^{(k)}_n = \exp[-i \hat{H}_{kn}
+dt_n]``, where ``\hat{H}_{kn}`` is the operator obtained by
+evaluating `trajectories[k].generator` on the ``n``'th time interval.
+
+The backward-propagation of ``|\chi_k(t)⟩`` has the boundary condition
+
+```math
+    |\chi_k(T)⟩ \equiv - \frac{\partial J_T}{\partial ⟨\Psi_k(T)|}\,.
+```
+
+The final-time gradient ``\nabla J_T`` is combined with the gradient for the
+running costs, and the total gradient is then fed into an optimizer
+(L-BFGS-B by default) that iteratively changes the values ``\{ϵ_{nl}\}`` to
+minimize ``J``.
+
+See [Background](@ref GRAPE-Background) for details.
+
+Returns a [`GrapeResult`](@ref).
+
+# Positional arguments
+
+* `trajectories`: A vector of [`Trajectory`](@extref `QuantumControl.Trajectory`)
+  objects. Each trajectory contains an `initial_state` and a dynamical
+  `generator` (e.g., time-dependent Hamiltonian). Each
+  trajectory may also contain arbitrary additional attributes like
+  `target_state` to be used in the `J_T` functional
+* `tlist`: A vector of time grid values.
+
+# Required keyword arguments
+
+* `J_T`: A function `J_T(Ψ, trajectories)` that evaluates the final time
+  functional from a list `Ψ` of forward-propagated states and
+  `trajectories`. The function `J_T` may also take a keyword argument
[…]
