+++ previews/PR93/refs.bib
@@ -1,0 +1,753 @@
+@string{aamop = "Adv. At. Mol. Opt. Phys."}
+@string{aarc = "Autom. Rem. Contr."}
+@string{ac = "Anal. Chem."}
+@string{acie = "Angew. Chem. Int. Ed."}
+@string{aipa = "AIP Advances"}
+@string{ajp = "Am. J. Phys."}
+@string{algo = "Algorithmica"}
+@string{ao = "Appl. Opt."}
+@string{ap = "Adv. Phys."}
+@string{apb = "Appl. Phys. B"}
+@string{apl = "Appl. Phys. Lett."}
+@string{apx = "Adv. Phys. X"}
+@string{aqt = "Adv. Quantum Tech."}
+@string{arcmp = "Annu. Rev. Condens. Matter Phys."}
+@string{arpc = "Annu. Rev. Phys. Chem."}
+@string{astrocomp = "Astron. Comput."}
+@string{atms = "ACM Trans. Math. Softw."}
+@string{avsqs = "AVS Quantum Sci."}
+@string{bit = "BIT"}
+@string{bstj = "Bell System Tech. J."}
+@string{cacm = "Commun. ACM"}
+@string{ccyb = "Control Cybern."}
+@string{cmp = "Commun. Math. Phys."}
+@string{computj = "Comput. J."}
+@string{contp = "Contemp. Phys."}
+@string{cp = "Chem. Phys."}
+@string{cpam = "Commun. Pur. Appl. Math."}
+@string{cpc = "Comput. Phys. Commun."}
+@string{cpl = "Chem. Phys. Lett."}
+@string{cse = "Comput. Sci. Eng."}
+@string{csj = "CEAS Space J."}
+@string{ecyb = "Engrg. Cybernetics"}
+@string{ejc = "Eur. J. Control"}
+@string{ejp = "Eur. J. Phys."}
+@string{electr = "Electronics"}
+@string{entr = "Entropy"}
+@string{epjb = "Eur. Phys. J. B"}
+@string{epjd = "Eur. Phys. J. D"}
+@string{epjp = "Eur. Phys. J. Plus"}
+@string{epjqt = "EPJ Quantum Technol."}
+@string{epl = "Europhys. Lett."}
+@string{farad = "Faraday Disc."}
+@string{foundphys = "Found. Phys."}
+@string{fp = "Fortschr. Phys."}
+@string{icta = "IET Control Theory Appl."}
+@string{ijqe = "IEEE J. Quantum Electron."}
+@string{ijqi = "Int. J. Quantum Inform."}
+@string{ijtp = "Int. J. Theor. Phys."}
+@string{imajam = "IMA J. Appl. Math."}
+@string{ip = "Inverse Problems"}
+@string{itac = "IEEE Trans. Automat. Contr."}
+@string{itas = "IEEE Trans. on Appl. Superc."}
+@string{jap = "J. Appl. Phys."}
+@string{jcam = "J. Comput. Appl. Math"}
+@string{jcmpp = "J. Comput. Phys."}
+@string{jcp = "J. Chem. Phys."}
+@string{jcpm = "J. Phys. Condens. Matter"}
+@string{jcss = "J. Comput. System Sci."}
+@string{jctn = "J. Comput. Theor. Nanos."}
+@string{jlum = "J. Lumin."}
+@string{jmo = "J. Mod. Opt."}
+@string{jmp = "J. Math. Phys."}
+@string{jmr = "J. Magnet. Res."}
+@string{jmra = "J. Magnet. Res. A"}
+@string{job = "J. Optics B"}
+@string{jors = "J. Open Res. Softw."}
[…]
