### 2025-10-01 e75d2d8
Add minimal example to the README
