+++ README.md
@@ -16,9 +16,6 @@ SPDX-License-Identifier: CC-BY-4.0
 
 Implementation of GRadient Ascent Pulse Engineering (GRAPE)
 
-Part of the [JuliaQuantumControl] organization, to be used in conjunction with the [`QuantumControl.jl`] framework.
-
-
 ## Installation
 
 [As usual for a registered Julia package](https://docs.julialang.org/en/v1/stdlib/Pkg/), `GRAPE` can be installed by typing
@@ -29,24 +26,67 @@ Part of the [JuliaQuantumControl] organization, to be used in conjunction with t
 
 in the Julia REPL.
 
-## Usage
+## Usage Example
+
+A minimal working example optimizing a state-to-state transition `|0⟩ → |1⟩` in a two-level quantum system:
 
 ```julia
-using QuantumControl
 using GRAPE
 
-# Set up a `QuantumControl.ControlProblem`
+using QuantumControl: hamiltonian  # data structure for `H = H₀ + ϵ(t) H₁`
+using QuantumControl.Functionals: J_T_sm  # square-modulus functional
+using QuantumPropagators: ExpProp  # propagation method: matrix exponentiation
 
-result = QuantumControl.optimize(problem; method=GRAPE)
+ϵ(t) = 0.2 # guess pulse
+
+H = hamiltonian([1  0; 0 -1], ([0  1; 1  0], ϵ))  # time-dependent Hamiltonian
+ket_0, ket_1 = ComplexF64[1, 0], ComplexF64[0, 1]  # basis states |0⟩, |1⟩
+tlist = collect(range(0, 5, length=501));  # time grid; final time T = 5.0
+
+# Optimization functionals depend on states |Ψ(T)⟩, described by a "trajectory"
+traj = GRAPE.Trajectory(
+    initial_state = ket_0,
+    generator = H,
+    target_state = ket_1
+);
+
+result = GRAPE.optimize(
+    [traj], tlist;
+    prop_method = ExpProp,  # suitable for small systems only!
+    J_T = J_T_sm,  #  J_T = 1 - |⟨Ψ(T)|1⟩|²
+    check_convergence=res -> begin
+        # without convergence check, stop after 5000 iterations
+        ((res.J_T < 1e-3) && (res.converged = true) && (res.message = "J_T < 10⁻³"))
+    end,
+)
+
+ϵ_opt = result.optimized_controls[1]
+
+
+# Or, using the QuantumControl API (recommended)
+
+using QuantumControl: ControlProblem, optimize, @optimize_or_load
+
+problem = ControlProblem(
+    [traj], tlist,
+    prop_method = ExpProp,
+    J_T = J_T_sm,
+    check_convergence=res -> begin
+        ((res.J_T < 1e-3) && (res.converged = true) && (res.message = "J_T < 10⁻³"))
+    end,
+)
+
+result = optimize(problem; method=GRAPE)
+
+# This dumps the optimization result in `tls_opt.jld2`
+result = @optimize_or_load("tls_opt.jld2", problem; method = GRAPE)
 ```
 
-See the [Usage section in the Documentation](https://juliaquantumcontrol.github.io/GRAPE.jl/stable/usage/)
+See the [Tutorial](https://juliaquantumcontrol.github.io/GRAPE.jl/stable/tutorial/) and [Usage section](https://juliaquantumcontrol.github.io/GRAPE.jl/stable/usage/) in the documentation for more details.
 
 ## Documentation
 
 The documentation of `GRAPE.jl` is available at <https://juliaquantumcontrol.github.io/GRAPE.jl>.
-
-For a broader perspective, also see the [documentation of the `QuantumControl.jl` framework](https://juliaquantumcontrol.github.io/QuantumControl.jl/).
 
 ## Contributing
 
