### 2025-10-01 a7f4cc7
Release 0.8.0
