+++ Project.toml
@@ -1,7 +1,7 @@
 name = "GRAPE"
 uuid = "6b52fcaf-80fe-489a-93e9-9f92080510be"
 authors = ["Michael Goerz <mail@michaelgoerz.net>"]
-version = "0.7.6+dev"
+version = "0.8.0"
 
 [deps]
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
@@ -24,7 +24,7 @@ LBFGSB = "0.4"
 LinearAlgebra = "1"
 Optim = "1"
 Printf = "1"
-QuantumControl = ">=0.11.1"
-QuantumGradientGenerators = ">=0.1.8"
+QuantumControl = "0.11.1"
+QuantumGradientGenerators = "0.1.8"
 SciMLPublic = "1.0.0"
 julia = "1.9"
