### 2025-10-02 #8813 @goerz
@AlCap23 @szabo137 Again, thank you for your comments!

> I think a simple quickstart-section in the documentation of GRAPE.jl would help understanding its usage, even if most of the setup comes from QuantumControl.jl.

Agreed: we have added both a ["minimum working example"](https://github.com/JuliaQuantumControl/GRAPE.jl?tab=readme-ov-file#usage-example)  to the README and a longer [Tutorial](https://juliaquantumcontrol.github.io/GRAPE.jl/stable/tutorial/) to the main documentation.

> GRAPE.jl look like a one-function extension of the functionality of QuantumControl.jl

In the original submission, we had decided to hide the API of the `GRAPE` package entirely behind `QuantumControl`, exploiting the somewhat unique multiple-dispatch paradigm of the Julia language. However, your comment shows the potential for confusion stemming from this choice, and shows that we should be much more explicit in the API design. We have refactored the `GRAPE` package slightly and now [define a standalone API](https://juliaquantumcontrol.github.io/GRAPE.jl/stable/api/#API).

We have also clarified the [Relation to the QuantumControl Framework](https://juliaquantumcontrol.github.io/GRAPE.jl/stable/usage/#Relation-to-the-QuantumControl-Framework) in the documentation explicitly, since using `GRAPE` via the `QuantumControl` API is still possible and even encouraged.

The `QuantumControl` package still provides useful building blocks for defining quantum control problems in general, whether or not these problems are then solved via `GRAPE`. In principle, one could forgo using any of these building blocks from `QuantumControl`, by using custom data structures for time-dependent Hamiltonians or defining custom optimization functionals. Quantum control packages outside of our own organization, such as [PRONTO](https://narijauskas.github.io/PRONTO.jl/dev/) and [Piccolo](https://www.harmoniqs.co), would be less likely to use these building blocks, even if they choose to plug in to the `QuantumControl.optimize` API. We have been in discussions with the authors of these packages about the possibility of integrating into the `QuantumControl` API. This would provide the ability to easily switch between control methods, and to compare their relative benefits for a given physical problem.

In any case, I want to stress that the `GRAPE.optimize_grape` implements the entire optimization in a self-contained way. I would consider this implementation rather non-trivial, both [conceptually](https://juliaquantumcontrol.github.io/GRAPE.jl/dev/background/) and in terms of [lines of code](https://github.com/openjournals/joss-reviews/issues/8813#issuecomment-3188117054), and well above JOSS' [criteria for substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort).

@Nikoleta-v3

> As I understand it, QuantumControl.jl serves as a high-level orchestration framework, while GRAPE.jl implements a specific optimization algorithm. The two can evolve independently (i.e., GRAPE.jl can update without requiring changes to QuantumControl.jl, and vice versa).

That is correct.

> The documentation suggests that users can choose between optimization methods and install only what they need (e.g., someone using only Krotov’s method doesn’t need GRAPE.jl).

Yes, that is correct. This also extends to third-party packages implementing other quantum control methods if they decide to be compatible with the `QuantumControl` framework (or we decide to write a wrapper), as well as other methods we are planning to implement ourselves in the future, beyond GRAPE and Krotov.

The relation between `QuantumControl` and `GRAPE`/`Krotov` is similar to that of [`Optimization.jl`](https://github.com/SciML/Optimization.jl) and its "solver" packages, like [`BlackBoxOptim.jl`](https://github.com/robertfeldt/BlackBoxOptim.jl), [`Manopt.jl`](https://manoptjl.org/stable/), etc.

> In that case, should Krotov.jl also be published independently under the same rationale?

Yes, absolutely. We would probably plan to submit a JOSS paper for Krotov within the next few months. There is still a lot of work left in Krotov's documentation before that can happen, in particular the equivalent to GRAPE's [background](https://juliaquantumcontrol.github.io/GRAPE.jl/stable/background/).

> Has QuantumControl.jl itself been published (as far as I can tell it hasn't) or is there a publication planned?

It has not. We'd probably plan some kind of publication _eventually_, but I'm not sure yet what exactly the scope of that publication would be. The rationale for `QuantumControl.jl` by itself would basically be "Provides a common interface for potentially all methods of quantum control, allowing easy comparisons between different methods". I could potentially see a longer "review paper" as the best way to publish that. Not sure yet.

> More broadly, why should GRAPE.jl be the subject of a single JOSS paper, rather than the framework and ecosystem together?

The ecosystem as a whole is a very open-ended endeavor that would never reach a point of completion where it would be "publishable". `GRAPE` and `Krotov` are certainly not the only two methods that would be part of the ecosystem – not even the methods we would be interested in implementing ourselves in the long term.

In contrast, `QuantumControl` as a framework and wrapper is well-defined in a much narrower sense (even if it could evolve to accommodate broader needs) of just providing a collective API.

The solver packages really are quite independent, and generally are pretty substantial projects in themselves (both GRAPE and Krotov have an entire community behind them, with several generations of refinements and variations; as do some other slightly newer methods like CRAB, and GOAT/GROUP). I would certainly feel that any future implementation of these other methods (or methods that may still be developed) would be more than justified in publishing their implementation as a JOSS paper, even if they adopt the `QuantumControl` API or make use of some of the building blocks that the framework provides.
