### 2025-10-02 fa607f3
Use QuantumPropagators in standalone examples
