### 2025-10-02 f730327
Allow `convergence_check` to return Bool or String
This allows for for more concise `convergence_check` implementations
that do not require / encourage mutating `results`.
