+++ previews/PR94/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"references/#References","page":"References","title":"References","text":"","category":"section"},{"location":"references/","page":"References","title":"References","text":"N. Khaneja, T. Reiss, C. Kehlet, T. Schulte-Herbrüggen and S. J. Glaser. Optimal control of coupled spin dynamics: design of NMR pulse sequences by gradient ascent algorithms. J. Magnet. Res. 172, 296 (2005).\n\n\n\nP. de Fouquières, S. G. Schirmer, S. J. Glaser and I. Kuprov. Second order gradient ascent pulse engineering. J. Magnet. Res. 212, 412 (2011).\n\n\n\nM. H. Goerz, S. C. Carrasco and V. S. Malinovsky. Quantum Optimal Control via Semi-Automatic Differentiation. Quantum 6, 871 (2022).\n\n\n\nH. Tal-Ezer and R. Kosloff. An Accurate and Efficient Scheme for Propagating the Time Dependent Schrödinger Equation. J. Chem. Phys. 81, 3967 (1984).\n\n\n\nD. L. Goodwin and I. Kuprov. Auxiliary matrix formalism for interaction representation transformations, optimal control, and spin relaxation theories. J. Chem. Phys. 143, 084113 (2015).\n\n\n\nM. H. Goerz, D. Basilewitsch, F. Gago-Encinas, M. G. Krauss, K. P. Horn, D. M. Reich and C. P. Koch. Krotov: A Python implementation of Krotov's method for quantum optimal control. SciPost Phys. 7, 080 (2019).\n\n\n\nZ. Tošner, T. Vosegaard, C. Kehlet, N. Khaneja, S. J. Glaser and N. C. Nielsen. Optimal control in NMR spectroscopy: Numerical implementation in SIMPSON. J. Magnet. Res. 197, 120 (2009).\n\n\n\nH. J. Hogben, M. Krzystyniak, G. T. Charnock, P. J. Hore and I. Kuprov. Spinach – A software library for simulation of spin dynamics in large spin systems. J. Magnet. Res. 208, 179 (2011).\n\n\n\nC. A. Ryan and contributors, pulse-finder: Matlab code for GRAPE optimal control in NMR (2013).\n\n\n\nJ. R. Johansson, P. D. Nation and F. Nori. QuTiP 2: A Python framework for the dynamics of open quantum systems. Comput. Phys. Commun. 184, 1234 (2013).\n\n\n\nN. Wittler, F. Roy, K. Pack, M. Werninghaus, A. S. Roy, D. J. Egger, S. Filipp, F. K. Wilhelm and S. Machnes. Integrated Tool Set for Control, Calibration, and Characterization of Quantum Devices Applied to Superconducting Qubits. Phys. Rev. Applied 15, 034080 (2021).\n\n\n\nM. Rossignolo, T. Reisser, A. Marshall, P. Rembold, A. Pagano, P. J. Vetter, R. S. Said, M. M. Müller, F. Motzoi, T. Calarco, F. Jelezko and S. Montangero. QuOCS: The quantum optimal control suite. Comput. Phys. Commun. 291, 108782 (2023).\n\n\n\nM. Zhang, H.-M. Yu, H. Yuan, X. Wang, R. Demkowicz-Dobrzański and J. Liu. QuanEstimation: An open-source toolkit for quantum parameter estimation. Phys. Rev. Research 4, 043057 (2022).\n\n\n\nQDYN: Fortran 95 library and utilities for quantum dynamics and optimal control (2025).\n\n\n\nC. Rackauckas and Q. Nie. DifferentialEquations.jl – A Performant
[…]
