### 2025-10-20 #8813 @Nikoleta-v3
Thank you for your reply, @goerz! Then I believe that `GRAPE.jl` meets the criteria for publication in JOSS. Thank you also for adding more details to the documentation.

@AlCap23 @szabo137, did you get a chance to look over the reply?
