+++ .github/workflows/JOSS.yml
@@ -20,7 +20,7 @@ jobs:
             openjournals/inara:latest \
             -o pdf -p paper.md
       - name: Upload
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v5
         with:
           name: paper
           path: paper/paper.pdf
