+++ test/test_iterations.jl
@@ -20,16 +20,16 @@ PASSTHROUGH = false
     # optimization to those numbers
     rng = StableRNG(1244568944)
     problem = dummy_control_problem(;
-        iter_start=10,
-        N=2,
-        density=1.0,
-        complex_operators=false,
+        iter_start = 10,
+        N = 2,
+        density = 1.0,
+        complex_operators = false,
         rng,
-        J_T=J_T_ss,
-        store_iter_info=["iter.", "J_T"]
+        J_T = J_T_ss,
+        store_iter_info = ["iter.", "J_T"]
     )
-    captured = IOCapture.capture(passthrough=PASSTHROUGH) do
-        optimize(problem; method=GRAPE, iter_stop=12)
+    captured = IOCapture.capture(passthrough = PASSTHROUGH) do
+        optimize(problem; method = GRAPE, iter_stop = 12)
     end
     res = captured.value
     @test res.converged
@@ -62,23 +62,23 @@ end
     end
 
     problem = dummy_control_problem(;
-        N=2,
-        density=1.0,
-        complex_operators=false,
+        N = 2,
+        density = 1.0,
+        complex_operators = false,
         rng,
-        J_T=J_T_ss,
-        callback=callback1,
+        J_T = J_T_ss,
+        callback = callback1,
     )
 
-    captured = IOCapture.capture(passthrough=PASSTHROUGH) do
-        optimize(problem; method=GRAPE, iter_stop=1)
+    captured = IOCapture.capture(passthrough = PASSTHROUGH) do
+        optimize(problem; method = GRAPE, iter_stop = 1)
     end
     @test contains(captured.output, "This is callback 1 for iter 0\n iter. ")
     @test contains(captured.output, "This is callback 1 for iter 1\n     1")
 
     # passing `callback` to `optimize` overwrites `callback` in `problem`
-    captured = IOCapture.capture(passthrough=PASSTHROUGH) do
-        optimize(problem; method=GRAPE, iter_stop=1, callback=callback2)
+    captured = IOCapture.capture(passthrough = PASSTHROUGH) do
+        optimize(problem; method = GRAPE, iter_stop = 1, callback = callback2)
     end
     @test captured.value.converged
     @test !contains(captured.output, "This is callback 1 for iter 0")
@@ -86,13 +86,13 @@ end
     @test contains(captured.output, "This is callback 2 for iter 0")
     @test contains(captured.output, "This is callback 2 for iter 1")
 
-    captured = IOCapture.capture(passthrough=PASSTHROUGH) do
+    captured = IOCapture.capture(passthrough = PASSTHROUGH) do
         optimize(
             problem;
-            method=GRAPE,
-            iter_stop=1,
-            callback=(callback1, callback2),
-            print_iters=false
+            method = GRAPE,
+            iter_stop = 1,
+            callback = (callback1, callback2),
+            print_iters = false
         )
     end
     @test captured.value.converged
@@ -107,13 +107,13 @@ end
     )
     @test captured.value.records == [("cb2", 0), ("cb2", 1)]
 
-    captured = IOCapture.capture(passthrough=PASSTHROUGH) do
+    captured = IOCapture.capture(passthrough = PASSTHROUGH) do
         optimize(
             problem;
-            method=GRAPE,
-            iter_stop=1,
-            callback=(callback1, callback2),
-            store_iter_info=["J_T"]
+            method = GRAPE,
+            iter_stop = 1,
+            callback = (callback1, callback2),
+            store_iter_info = ["J_T"]
         )
     end
     @test captured.value.converged
@@ -125,13 +125,13 @@ end
 
     # we should also be able to modify the updated pulses in the callback and
[…]
