### 2025-10-30 #8813 @goerz
@Nikoleta-v3 I've tagged the [`v1.0.0` release](https://github.com/JuliaQuantumControl/GRAPE.jl/releases/tag/v1.0.0). The [Zenodo snapshot](https://zenodo.org/records/17487150) for this release has the DOI [10.5281/zenodo.17487150](https://doi.org/10.5281/zenodo.17487150)
