### 2025-11-01 #8813 @goerz
@Nikoleta-v3

> Could you update the authors’ names on Zenodo to include the middle-name character/initial?

Done
